refactor: commons 디렉토리 구조 변경
공용 컴포넌트의 경우, 타입을 따로 빼는 것이 불필요하다고 생각하여 type.d.ts 파일을 따로 만들지 않았습니다.
