+++ src/commons/components/UserToggleBox.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { getCookie, removeToken } from './cookie/cookie';
+import { getCookie, removeToken } from '../cookie/cookie';
 
 const UserToggleBox = () => {
   const token = getCookie('loginToken');
