+++ src/pages/signUp/components/VSignupInputForm.tsx
@@ -1,5 +1,5 @@
 import AddressInputGroup from 'pages/signUp/components/AddressInputGroup';
-import InputGroup from 'commons/InputGroup';
+import InputGroup from 'commons/components/InputGroup';
 import { ClipLoader } from 'react-spinners';
 import { VSignupInputProps } from '../signupType';
 import ValidateText from './ValidateText';
